backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.17
